+++ crates/fosr/src/inject/ebpf.rs
@@ -1,4 +1,4 @@
-use crate::stage4::*;
+use crate::inject::*;
 use aya::Ebpf;
 use pnet::datalink;
 
@@ -17,6 +17,9 @@ impl NetEnabler for EBPFNetEnabler {
     }
     fn close_session(&self, _: &FlowId) {}
     fn open_session(&self, _: &FlowId) {}
+    fn get_ttl(&self) -> Option<u8> {
+        None
+    }
 }
 
 /// Load the ebpf XDP program, and attach it to each valid interface used by Fos-R.
@@ -30,11 +33,8 @@ fn load_ebpf_program(local_interfaces: &[datalink::NetworkInterface]) {
     // Retrieve the stored eBPF program that where stored, at compilation, into the binary
     // This only hold a reference, the object is stored by aya (globaly), so no need to store it anywhere,
     // it will not be destroyed at the end of the function
-    let mut ebpf = aya::Ebpf::load(aya::include_bytes_aligned!(concat!(
-        env!("OUT_DIR"),
-        "/fosr-ebpf-prog"
-    )))
-    .expect("Couldn't retrieve eBPF program");
+    let mut ebpf =
+        aya::Ebpf::load(fosr_ebpf::EBPF_PROGRAM).expect("Couldn't retrieve eBPF program");
     let program: &mut Xdp = ebpf
         .program_mut("fosr_ebpf")
         .expect("Failed to get mut reference of program")