+++ crates/fosr/src/inject/mod.rs
@@ -1,10 +1,12 @@
 use crate::structs::*;
 
 #[cfg(feature = "net_injection")]
+/// Fast injection, that does not wait for packet response
 pub mod fast;
 #[cfg(feature = "net_injection")]
 pub use fast::start_fast;
 #[cfg(feature = "net_injection")]
+/// Reliable injection, that waits for packet response
 pub mod reliable;
 #[cfg(feature = "net_injection")]
 pub use reliable::start_reliable;
@@ -14,26 +16,31 @@ pub use reliable::start_reliable;
     feature = "ebpf",
     feature = "net_injection"
 ))]
+/// Network enabler based on eBPF
 pub mod ebpf;
 #[cfg(all(target_os = "linux", feature = "iptables", feature = "net_injection"))]
+/// Network enable based on iptables
 pub mod iptables;
 
+/// A trait for network enablers
 pub trait NetEnabler: Clone + std::marker::Send + 'static {
-    // is this packet sent by Fos-R ?
+    fn get_ttl(&self) -> Option<u8>;
+
+    /// is this packet sent by Fos-R ?
     fn is_packet_relevant(&self, flags: u8) -> bool;
 
-    // should we send the packet without waiting for any answer?
+    /// should we send the packet without waiting for any answer?
     fn is_fast(&self) -> bool;
 
-    // close the connection
+    /// close the connection
     fn close_session(&self, f: &FlowId);
 
-    // set-up the connection
+    /// set-up the connection
     fn open_session(&self, f: &FlowId);
 }
 
 #[derive(Debug, Clone)]
-// the dummy net enabler is used when no network injection is performed
+/// Used when no network injection is performed
 pub struct DummyNetEnabler {}
 
 impl NetEnabler for DummyNetEnabler {
@@ -45,4 +52,7 @@ impl NetEnabler for DummyNetEnabler {
     }
     fn close_session(&self, _: &FlowId) {}
     fn open_session(&self, _: &FlowId) {}
+    fn get_ttl(&self) -> Option<u8> {
+        None
+    }
 }