+++ crates/fosr/src/main.rs
@@ -1,24 +1,27 @@
 // we access the code through the library
+#[cfg(feature = "net_injection")]
+use fosr::inject;
 use fosr::pcap2flow;
 use fosr::stage0;
 use fosr::stage1;
 use fosr::stage2;
 use fosr::stage3;
-#[cfg(feature = "net_injection")]
-use fosr::stage4;
 use fosr::structs::*;
-use fosr::ui::Target;
+use fosr::stats::Target;
 use fosr::*;
 mod cmd; // cmd is not part of the library
 
 use std::cmp::max;
 use std::collections::HashMap;
 use std::fs;
+use std::fs::OpenOptions;
+use std::io::BufWriter;
 use std::net::Ipv4Addr;
 use std::path::Path;
 use std::process;
 use std::sync::Arc;
 use std::thread;
+use std::time::Instant;
 use std::time::{Duration, SystemTime, UNIX_EPOCH};
 
 use chrono::DateTime;
@@ -27,6 +30,9 @@ use chrono::TimeZone;
 use chrono_tz::Tz;
 use clap::Parser;
 use crossbeam_channel::bounded;
+use indicatif::HumanBytes;
+use pcap_file::pcap::{PcapPacket, PcapWriter};
+#[cfg(feature = "net_injection")]
 use pnet::{datalink, ipnetwork::IpNetwork};
 
 const CHANNEL_SIZE: usize = 50;
@@ -38,7 +44,7 @@ struct Profile {
     config: config::Hosts,
 }
 
-struct S4Param<T: stage4::NetEnabler> {
+struct InjectParam<T: inject::NetEnabler> {
     net_enabler: T,
     injection_algo: cmd::InjectionAlgo,
 }
@@ -89,32 +95,6 @@ fn main() {
     stage1::bayesian_networks::BNGenerator::test();
     let args = cmd::Args::parse();
 
-    // Extract all IPv4 local interfaces (except loopback)
-    let extract_addr = |iface: datalink::NetworkInterface| {
-        iface
-            .ips
-            .into_iter()
-            .filter(IpNetwork::is_ipv4)
-            .map(|i| match i {
-                IpNetwork::V4(data) => data.ip(),
-                _ => unreachable!(),
-            })
-    };
-    // the local interfaces are used by the stage 4 and to identify local IPs
-    // we do not include loopback interfaces or interfaces without an IPv4 address
-    let local_interfaces: Vec<datalink::NetworkInterface> = datalink::interfaces()
-        .into_iter()
-        .filter(|iface| !iface.is_loopback() && iface.ips.iter().any(IpNetwork::is_ipv4))
-        .collect();
-    // for each interface, we extract its addresses
-    let local_ips: Vec<Ipv4Addr> = local_interfaces
-        .clone()
-        .into_iter()
-        .flat_map(extract_addr)
-        .filter(|i| !i.is_loopback())
-        .collect();
-    log::debug!("IPv4 interfaces: {:?}", &local_ips);
-
     match args.command {
         cmd::Command::Pcap2Flow {
             input_pcap,
@@ -140,6 +120,33 @@ fn main() {
         } => {
             #[cfg(not(all(target_os = "linux", feature = "iptables")))]
             let stealthy = false;
+
+            // Extract all IPv4 local interfaces (except loopback)
+            let extract_addr = |iface: datalink::NetworkInterface| {
+                iface
+                    .ips
+                    .into_iter()
+                    .filter(IpNetwork::is_ipv4)
+                    .map(|i| match i {
+                        IpNetwork::V4(data) => data.ip(),
+                        _ => unreachable!(),
+                    })
+            };
+            // the local interfaces are used by the inject module and used to identify local IPs
+            // we do not include loopback interfaces or interfaces without an IPv4 address
+            let local_interfaces: Vec<datalink::NetworkInterface> = datalink::interfaces()
+                .into_iter()
+                .filter(|iface| !iface.is_loopback() && iface.ips.iter().any(IpNetwork::is_ipv4))
+                .collect();
+
+            // for each interface, we extract its addresses
+            let local_ips: Vec<Ipv4Addr> = local_interfaces
+                .clone()
+                .into_iter()
+                .flat_map(extract_addr)
+                .filter(|i| !i.is_loopback())
+                .collect();
+            log::debug!("IPv4 interfaces: {:?}", &local_ips);
 
             // identify the role of the current host
             let profile = Profile::load(profile.as_deref());
@@ -183,8 +190,8 @@ fn main() {
             match net_enabler {
                 #[cfg(all(any(target_os = "windows", target_os = "linux"), feature = "ebpf"))]
                 cmd::NetEnabler::Ebpf => {
-                    let s4net = S4Param {
-                        net_enabler: stage4::ebpf::EBPFNetEnabler::new(false, &local_interfaces),
+                    let s4net = InjectParam {
+                        net_enabler: inject::ebpf::EBPFNetEnabler::new(false, &local_interfaces),
                         injection_algo,
                     };
                     run(
@@ -197,14 +204,14 @@ fn main() {
                         (s1, 1),
                         (s2, 1),
                         (s3, 1),
-                        Arc::new(ui::Stats::default()),
+                        Arc::new(stats::Stats::default()),
                         Some(s4net),
                     );
                 }
                 #[cfg(all(target_os = "linux", feature = "iptables"))]
                 cmd::NetEnabler::Iptables => {
-                    let s4net = S4Param {
-                        net_enabler: stage4::iptables::IPTablesNetEnabler::new(!stealthy, false),
+                    let s4net = InjectParam {
+                        net_enabler: inject::iptables::IPTablesNetEnabler::new(!stealthy, false),
                         injection_algo,
                     };
                     run(
@@ -217,7 +224,7 @@ fn main() {
                         (s1, 1),
                         (s2, 1),
                         (s3, 1),
-                        Arc::new(ui::Stats::default()),
+                        Arc::new(stats::Stats::default()),
                         Some(s4net),
                     );
                 }
@@ -228,12 +235,13 @@ fn main() {
             profile,
             outfile,
             packets_count,
-            minimum_threads,
+            monothread,
             order_pcap,
             start_time,
             duration,
             flow_per_day,
             tz,
+            taint,
         } => {
             // load the models
             let profile = Profile::load(profile.as_deref());
@@ -320,93 +328,48 @@ fn main() {
             let s1 =
                 stage1::flowchronicle::FCGenerator::new(patterns, profile.config.clone(), false);
             let s2 = stage2::tadam::TadamGenerator::new(automata_library);
-            let s3 = stage3::Stage3::new(false, profile.config);
-
-            let (s1_count, s2_count, s3_count) = if minimum_threads {
-                (1, 1, 1) // only use on thread per task
+            let s3 = stage3::Stage3::new(taint, profile.config);
+            if monothread {
+                log::info!("Monothread generation");
+                run_monothread(
+                    ExportParams {
+                        outfile,
+                        order_pcap,
+                    },
+                    s0,
+                    s1,
+                    s2,
+                    s3,
+                );
             } else {
                 let cpu_count = num_cpus::get();
-                (
+                if cpu_count <= 2 {
+                    log::warn!("This host has only {cpu_count} core(s). Consider using the --monothread option.");
+                }
+                let (s1_count, s2_count, s3_count) = (
                     max(1, cpu_count / 2),
                     max(1, cpu_count / 2),
                     max(1, cpu_count / 2),
-                ) // the total is indeed larger than cpu_count. This has been empirically assessed to be a correct heuristic to maximise the performances
-            };
-            run(
-                vec![],
-                Some(ExportParams {
-                    outfile,
-                    order_pcap,
-                }),
-                s0,
-                (s1, s1_count),
-                (s2, s2_count),
-                (s3, s3_count),
-                Arc::new(ui::Stats::new(target)),
-                None::<S4Param<stage4::DummyNetEnabler>>,
-            );
+                ); // the total is indeed larger than cpu_count. This has been empirically assessed to be a correct heuristic to maximise the performances
+
+                run(
+                    vec![],
+                    Some(ExportParams {
+                        outfile,
+                        order_pcap,
+                    }),
+                    s0,
+                    (s1, s1_count),
+                    (s2, s2_count),
+                    (s3, s3_count),
+                    Arc::new(stats::Stats::new(target)),
+                    None::<InjectParam<inject::DummyNetEnabler>>,
+                );
+            }
         }
         cmd::Command::Untaint { input, output } => {
             pcap2flow::untaint_file(&input, &output);
-        } // #[cfg(feature = "replay")]
-          // cmd::Command::Replay {
-          //     file,
-          //     // config_path,
-          //     taint,
-          //     fast,
-          // } => {
-          //     // Read content of the file
-          //     log::debug!("Initialize stages");
-          //     let mut flow_router_tx = HashMap::new();
-          //     let mut stage_4_rx = HashMap::new();
-
-          //     for proto in Protocol::iter() {
-          //         let (tx, rx) = bounded::<Packets>(crate::CHANNEL_SIZE);
-          //         flow_router_tx.insert(proto, tx);
-          //         stage_4_rx.insert(proto, rx);
-          //     }
-          //     // let ip_replacement_map: HashMap<Ipv4Addr, Ipv4Addr> = if let Some(path) = config_path {
-          //     //     // read from config file
-          //     //     replay::parse_config(
-          //     //         &fs::read_to_string(path).expect("Cannot access the configuration file."),
-          //     //     )
-          //     // } else {
-          //     //     // no IP replacement
-          //     //     HashMap::new()
-          //     // };
-
-          //     let stage_replay = replay::Replay::new();
-          //     let flows = stage_replay.parse_flows(&file);
-
-          //     // Flow router
-          //     let thread_builder = thread::Builder::new().name("replay_flow_router".to_string());
-          //     let flow_router = thread_builder
-          //         .spawn(move || {
-          //             let mut sent_flows = 0;
-          //             for flow in flows {
-          //                 let proto = flow.flow.get_proto();
-
-          //                 let tx = flow_router_tx.get(&proto).expect("Unknown protocol");
-          //                 stage3::send_online(&local_interfaces, flow, tx);
-          //                 sent_flows += 1;
-          //             }
-
-          //             log::info!("Sent {} flows to be replayed", sent_flows);
-          //         })
-          //         .unwrap();
-
-          //     // Stage 4
-          //     let mut stage_4 = stage4::Stage4::new(taint, fast);
-          //     let thread_builder = thread::Builder::new().name("replay_stage4".to_owned());
-          //     let stage_4_thread = thread_builder
-          //         .spawn(move || {
-          //             stage_4.start(stage_4_rx);
-          //         })
-          //         .unwrap();
-
-          //     flow_router.join().unwrap();
-          //     stage_4_thread.join().unwrap();
-          // }
+        }
     };
 }
 
@@ -437,15 +400,15 @@ struct ExportParams {
 /// - `stats`: an Arc to a structure containing generation statistics
 /// - `s4net`: an optional network enable
 #[allow(clippy::too_many_arguments)]
-fn run<T: stage4::NetEnabler>(
+fn run<T: inject::NetEnabler>(
     local_interfaces: Vec<Ipv4Addr>,
     export: Option<ExportParams>,
     s0: impl stage0::Stage0,
     s1: (impl stage1::Stage1, usize),
     s2: (impl stage2::Stage2, usize),
     s3: (stage3::Stage3, usize),
-    stats: Arc<ui::Stats>,
-    s4net: Option<S4Param<T>>,
+    stats: Arc<stats::Stats>,
+    s4net: Option<InjectParam<T>>,
 ) {
     let (s1, s1_count) = s1;
     let (s2, s2_count) = s2;
@@ -501,7 +464,7 @@ fn run<T: stage4::NetEnabler>(
         gen_threads.push(
             builder
                 .spawn(move || {
-                    let _ = stage0::run(s0, tx_s0, stats_s0);
+                    let _ = stage0::run_channel(s0, tx_s0, stats_s0);
                 })
                 .unwrap(),
         );
@@ -517,7 +480,7 @@ fn run<T: stage4::NetEnabler>(
             gen_threads.push(
                 builder
                     .spawn(move || {
-                        let _ = stage1::run(s1, rx_s1, tx_s1, stats);
+                        let _ = stage1::run_channel(s1, rx_s1, tx_s1, stats);
                     })
                     .unwrap(),
             );
@@ -534,7 +497,7 @@ fn run<T: stage4::NetEnabler>(
             gen_threads.push(
                 builder
                     .spawn(move || {
-                        let _ = stage2::run(s2, rx_s2, tx_s2, stats);
+                        let _ = stage2::run_channel(s2, rx_s2, tx_s2, stats);
                     })
                     .unwrap(),
             );
@@ -561,7 +524,7 @@ fn run<T: stage4::NetEnabler>(
                         gen_threads.push(
                             builder
                                 .spawn(move || {
-                                    let _ = stage3::run(
+                                    let _ = stage3::run_channel(
                                         |f, p, a| s3.generate_tcp_packets(f, p, a),
                                         local_interfaces,
                                         rx_s3_tcp,
@@ -579,7 +542,7 @@ fn run<T: stage4::NetEnabler>(
                         gen_threads.push(
                             builder
                                 .spawn(move || {
-                                    let _ = stage3::run(
+                                    let _ = stage3::run_channel(
                                         |f, p, a| s3.generate_udp_packets(f, p, a),
                                         local_interfaces,
                                         rx_s3_udp,
@@ -597,7 +560,7 @@ fn run<T: stage4::NetEnabler>(
                         gen_threads.push(
                             builder
                                 .spawn(move || {
-                                    let _ = stage3::run(
+                                    let _ = stage3::run_channel(
                                         |f, p, a| s3.generate_icmp_packets(f, p, a),
                                         local_interfaces,
                                         rx_s3_icmp,
@@ -636,15 +599,15 @@ fn run<T: stage4::NetEnabler>(
         #[cfg(feature = "net_injection")]
         if let Some(s4net) = s4net {
             let stats = Arc::clone(&stats);
-            let builder = thread::Builder::new().name("Stage4".into());
+            let builder = thread::Builder::new().name("inject".into());
             gen_threads.push(
                 builder
                     .spawn(move || match s4net.injection_algo {
                         cmd::InjectionAlgo::Fast => {
-                            stage4::start_fast(s4net.net_enabler, rx_s4, stats)
+                            inject::start_fast(s4net.net_enabler, rx_s4, stats)
                         }
                         cmd::InjectionAlgo::Reliable => {
-                            stage4::start_reliable(s4net.net_enabler, rx_s4, stats)
+                            inject::start_reliable(s4net.net_enabler, rx_s4, stats)
                         }
                     })
                     .unwrap(),
@@ -655,7 +618,7 @@ fn run<T: stage4::NetEnabler>(
     {
         let stats = Arc::clone(&stats);
         let builder = thread::Builder::new().name("Monitoring".into());
-        threads.push(builder.spawn(move || ui::run(stats)).unwrap());
+        threads.push(builder.spawn(move || stats::run(stats)).unwrap());
     }
 
     // Wait for the generation threads to end
@@ -676,5 +639,70 @@ fn run<T: stage4::NetEnabler>(
     // Wait for the other threads to stop
     for thread in threads {
         thread.join().unwrap();
+    }
+}
+
+/// Run the generation with only one thread
+fn run_monothread(
+    export: ExportParams,
+    s0: impl stage0::Stage0,
+    s1: impl stage1::Stage1,
+    s2: impl stage2::Stage2,
+    s3: stage3::Stage3,
+) {
+    let start = Instant::now();
+
+    log::info!("Stage 0 generation");
+    let vec = stage0::run_vec(s0);
+    log::info!("Stage 1 generation");
+    let vec = stage1::run_vec(s1, vec);
+    log::info!("Stage 2 generation");
+    let vec = stage2::run_vec(s2, vec);
+
+    let mut all_packets = vec![];
+
+    log::info!("Stage 3 generation");
+    all_packets.append(&mut stage3::run_vec(
+        |f, p, a| s3.generate_udp_packets(f, p, a),
+        vec.udp,
+    ));
+    all_packets.append(&mut stage3::run_vec(
+        |f, p, a| s3.generate_tcp_packets(f, p, a),
+        vec.tcp,
+    ));
+    all_packets.append(&mut stage3::run_vec(
+        |f, p, a| s3.generate_icmp_packets(f, p, a),
+        vec.icmp,
+    ));
+
+    let gen_duration = start.elapsed().as_secs_f64();
+    let total_size = all_packets.iter().map(|p| p.data.len()).sum::<usize>() as u64;
+    log::info!(
+        "Generation throughput: {}/s",
+        HumanBytes(((total_size as f64) / gen_duration) as u64)
+    );
+
+    if export.order_pcap {
+        log::info!("Sorting the packets");
+        all_packets.sort_unstable();
+    }
+
+    let file_out = OpenOptions::new()
+        .write(true)
+        .create(true)
+        .truncate(true)
+        .open(&export.outfile)
+        .expect("Error opening or creating file");
+    let mut pcap_writer = PcapWriter::new(BufWriter::new(file_out)).expect("Error writing file");
+
+    log::info!("Pcap export");
+    for packet in all_packets.iter() {
+        pcap_writer
+            .write_packet(&PcapPacket::new(
+                packet.timestamp,
+                packet.data.len() as u32,
+                &packet.data,
+            ))
+            .unwrap();
     }
 }